refactor: relocate mis-placed infrastructure out of _core/
- Move _core/http_server.py → _mcp/transport/http_server.py
- Move _core/session_manager.py → _persistence/session_manager.py
- Extract RecoveryManager from _core/error_tracking.py → _persistence/recovery.py
- Keep ErrorTracker (pure statistics) in _core/
- Update all imports across codebase and tests
