+++ clang_index_mcp/_persistence/session_manager.py
